# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: api-tracker
  version: "1.0"
  description: API-tracker REST API. Base URL — https://api.apitracker.ru
servers:
  - url: https://api.apitracker.ru
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT access token (RS256, 15 min) or PAT (prefix `pat_`).
  schemas:
...
    TaskPermissions:
      type: object
      properties:
        edit_title: { type: boolean }
        edit_description: { type: boolean }
        edit_tags: { type: boolean }
        edit_blockers: { type: boolean }
        edit_assignee: { type: boolean }
        edit_status: { type: boolean }
        share: { type: boolean }
        delete_task: { type: boolean }
    TaskDirectAccess:
      type: object
      properties:
        id: { type: string, format: uuid }
        task_id: { type: string, format: uuid }
        grantee_user_id: { type: string, format: uuid, nullable: true }
        grantee_team_id: { type: string, format: uuid, nullable: true }
        granted_by: { type: string, format: uuid }
        permissions: { $ref: "#/components/schemas/TaskPermissions" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, task_id, granted_by, permissions, created_at, updated_at]
...
    ProjectTeamMember:
      type: object
      properties:
        team_id: { type: string, format: uuid }
        team_name: { type: string }
        permissions: { $ref: "#/components/schemas/ProjectPermissions" }
        is_frozen_in_project: { type: boolean }
        joined_at: { type: string, format: date-time }
      required: [team_id, team_name, permissions, is_frozen_in_project, joined_at]
...
security:
  - BearerAuth: []
paths:
...
  /tasks/{id}/accesses:
    get:
      summary: List task direct accesses
      operationId: listTaskAccesses
      tags: [TaskAccesses]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of direct accesses
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskDirectAccess"
                required: [items]

    post:
      summary: Grant direct access to task
      operationId: grantTaskAccess
      tags: [TaskAccesses]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                grantee_user_id: { type: string, format: uuid }
                grantee_team_id: { type: string, format: uuid }
                permissions: { $ref: "#/components/schemas/TaskPermissions" }
              required: [permissions]
      responses:
        "201":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDirectAccess"
        "409":
          description: Access already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{id}/accesses/{access_id}:
    patch:
      summary: Update task direct access permissions
      operationId: updateTaskAccess
      tags: [TaskAccesses]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: access_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissions: { $ref: "#/components/schemas/TaskPermissions" }
              required: [permissions]
      responses:
        "200":
          description: Access updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDirectAccess"
        "404":
          description: Access not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Revoke task direct access
      operationId: revokeTaskAccess
      tags: [TaskAccesses]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: access_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Access revoked
        "404":
          description: Access not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects/{id}/members:
    get:
      summary: List project user members
      operationId: listProjectMembers
      tags: [ProjectMembers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of project members
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProjectMember"
                required: [items]
  /projects/{id}/members/{user_id}:
    patch:
      summary: Update project member permissions
      operationId: updateProjectMember
      tags: [ProjectMembers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissions: { $ref: "#/components/schemas/ProjectPermissions" }
              required: [permissions]
      responses:
        "200":
          description: Member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Remove member from project
      operationId: removeProjectMember
      tags: [ProjectMembers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Member removed
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects/{id}/team-members:
    get:
      summary: List project team members
      operationId: listProjectTeamMembers
      tags: [ProjectMembers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of project team members
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProjectTeamMember"
                required: [items]
  /projects/{id}/team-members/{team_id}:
    patch:
      summary: Update project team member permissions
      operationId: updateProjectTeamMember
      tags: [ProjectMembers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: team_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissions: { $ref: "#/components/schemas/ProjectPermissions" }
              required: [permissions]
      responses:
        "200":
          description: Team member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectTeamMember"
        "404":
          description: Team member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Remove team from project
      operationId: removeProjectTeamMember
      tags: [ProjectMembers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: team_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Team member removed
        "404":
          description: Team member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects:
    get:
      summary: List own projects (cursor pagination)
      operationId: listProjects
      tags: [Projects]
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        "200":
          description: Paginated project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                  next_cursor:
                    type: string
                    nullable: true
                  total:
                    type: integer
                required: [items, total]

    post:
      summary: Create a project
      operationId: createProject
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
              required: [name]
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects/{id}:
    get:
      summary: Get project by ID
      operationId: getProject
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update project fields
      operationId: updateProject
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete project
      operationId: deleteProject
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Project deleted
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /teams:
    get:
      summary: List own teams (cursor pagination)
      operationId: listTeams
      tags: [Teams]
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        "200":
          description: Paginated team list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Team"
                  next_cursor:
                    type: string
                    nullable: true
                  total:
                    type: integer
                required: [items, total]

    post:
      summary: Create a team
      operationId: createTeam
      tags: [Teams]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
              required: [name]
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /teams/{id}:
    get:
      summary: Get team by ID
      operationId: getTeam
      tags: [Teams]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update team fields
      operationId: updateTeam
      tags: [Teams]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete team
      operationId: deleteTeam
      tags: [Teams]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Team deleted
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /teams/{id}/members:
    get:
      summary: List team members
      operationId: listTeamMembers
      tags: [TeamMembers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of team members
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TeamMember"
                required: [items]
  /teams/{id}/members/{user_id}:
    patch:
      summary: Update team member role
      operationId: updateTeamMemberRole
      tags: [TeamMembers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
              required: [role]
      responses:
        "200":
          description: Member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Remove member from team
      operationId: removeTeamMember
      tags: [TeamMembers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Member removed
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks:
    get:
      summary: List tasks (RSQL filter, cursor pagination, sorting)
      operationId: listTasks
      tags: [Tasks]
      parameters:
        - name: filter
          in: query
          schema:
            type: string
          description: RSQL filter expression
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
          description: "Sort field:direction (e.g. created_at:asc, updated_at:desc, title:asc)"
      responses:
        "200":
          description: Paginated task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  next_cursor:
                    type: string
                    nullable: true
                  total:
                    type: integer
                required: [items, total]

    post:
      summary: Create a task
      operationId: createTask
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [opened, progress, closed]
                assignee_id:
                  type: string
                  format: uuid
                  nullable: true
                tags:
                  type: array
                  items:
                    type: string
                project_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                blocking_task_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
              required: [title]
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "422":
          description: Tariff limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{id}:
    get:
      summary: Get task by ID
      operationId: getTask
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update task fields
      operationId: updateTask
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [opened, progress, closed]
                assignee_id:
                  type: string
                  format: uuid
                  nullable: true
                tags:
                  type: array
                  items:
                    type: string
                blocking_task_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete task
      operationId: deleteTask
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{id}/projects:
    post:
      summary: Attach task to project
      operationId: attachTaskProject
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id:
                  type: string
                  format: uuid
              required: [project_id]
      responses:
        "204":
          description: Attached
        "409":
          description: Already attached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{id}/projects/{project_id}:
    delete:
      summary: Detach task from project
      operationId: detachTaskProject
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Detached
        "404":
          description: Association not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects/{id}/tasks:
    get:
      summary: List project tasks (RSQL filter, cursor pagination, sorting)
      operationId: listProjectTasks
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: filter
          in: query
          schema:
            type: string
          description: RSQL filter expression
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
          description: "Sort field:direction (e.g. created_at:asc, updated_at:desc, title:asc)"
      responses:
        "200":
          description: Paginated task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  next_cursor:
                    type: string
                    nullable: true
                  total:
                    type: integer
                required: [items, total]
  /projects/{id}/invitations:
    post:
      summary: Create project invitation
      operationId: createProjectInvitation
      tags: [Invitations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invitee_user_id: { type: string, format: uuid }
                invitee_team_id: { type: string, format: uuid }
                permissions: { $ref: "#/components/schemas/ProjectPermissions" }
              required: [permissions]
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectInvitation"

    get:
      summary: List project invitations
      operationId: listProjectInvitations
      tags: [Invitations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of project invitations
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProjectInvitation"
                required: [items]
  /projects/{id}/invitations/{invitation_id}:
    delete:
      summary: Delete project invitation
      operationId: deleteProjectInvitation
      tags: [Invitations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: invitation_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Invitation deleted
  /invitations/projects/{invitation_id}/accept:
    post:
      summary: Accept project invitation
      operationId: acceptProjectInvitation
      tags: [Invitations]
      parameters:
        - name: invitation_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Invitation accepted
...